feat: Consider user permissions before executing CRUD actions
Also add some comments to code.

Resolves: #30

+++ catalog/items/views.py
@@ -1,7 +1,7 @@
 from copy import copy
 from flask import (
     Blueprint, render_template, abort, jsonify, request,
-    redirect, url_for, flash)
+    redirect, url_for, flash, session)
 from sqlalchemy import and_, not_
 from sqlalchemy.orm import joinedload
 
@@ -17,10 +17,13 @@ items_bp = Blueprint('items', __name__)
 def details(iid):
     item = Item.query.options(joinedload(Item.category)).get(iid)
     delete_form = DeleteItemForm()
+    app_user_id = session.get('app_user_id', '')
+    can_modify_item = app_user_id and (item.app_user_id == app_user_id)
 
     if item:
         return render_template(
-            'items/details.html', item=item, delete_form=delete_form)
+            'items/details.html', item=item, delete_form=delete_form,
+            can_modify_item=can_modify_item)
     else:
         abort(404)
 
@@ -37,6 +40,13 @@ def details_json(iid):
 
 @items_bp.route('/add/', methods=('GET', 'POST'))
 def add():
+    # check that a user is logged in
+    app_user_id = session.get('app_user_id', '')
+    if not app_user_id:
+        flash("Log in to create items", 'error')
+        return redirect(url_for('categories.index'))
+
+    # verify that category id is in URL and that category exists
     cid = ''
     try:
         cid = int(request.args.get('cid'))
@@ -46,19 +56,25 @@ def add():
         flash("Invalid category id ({})".format(cid), 'error')
         return redirect(url_for('categories.index'))
 
+    # create form
     form = ItemForm()
     form.category_id.data = cid
 
+    # validate form if POST request
     if form.validate_on_submit():
+        # get item properties
         item_props = copy(form.data)
         del item_props['csrf_token']
+        item_props['app_user_id'] = app_user_id
 
+        # check that item with same name does not exist
         item = Item.query.filter_by(name=item_props['name']).first()
         if item:
             flash("name: Item with name '{}' already exists".format(
                 item_props['name']), 'error')
             return render_template('items/add.html', form=form)
 
+        # create item
         try:
             item = Item(**item_props)
             db.session.add(item)
@@ -70,6 +86,7 @@ def add():
             flash("Item created successfully", 'info')
             return redirect(url_for('items.details', iid=item.id))
     else:
+        # flash form errors
         for field, errors in form.errors.items():
             for error in errors:
                 flash("{}: {}".format(field, error), 'error')
@@ -79,6 +96,7 @@ def add():
 
 @items_bp.route('/edit/<int:iid>', methods=('GET', 'POST'))
 def edit(iid):
+    # check item existence
     try:
         item = Item.query.filter_by(id=iid).one()
     except Exception as e:
@@ -86,8 +104,16 @@ def edit(iid):
         flash("Item with id ({}) does not exist".format(iid), 'error')
         return redirect(url_for('categories.index'))
 
+    # check that logged in user is the creator of the item
+    app_user_id = session.get('app_user_id', '')
+    if not (app_user_id and (item.app_user_id == app_user_id)):
+        flash("You can't edit this item. You are not its creator", 'error')
+        return redirect(url_for('items.details', iid=iid))
+
+    # create form
     form = ItemForm(obj=item)
 
+    # validate form if POST request
     if form.validate_on_submit():
         # check name duplicity
         same_name_item = Item.query.filter(
@@ -119,6 +145,7 @@ def edit(iid):
             flash("Item modified successfully", 'info')
             return redirect(url_for('items.details', iid=item.id))
     else:
+        # flash form errors
         for field, errors in form.errors.items():
             for error in errors:
                 flash("{}: {}".format(field, error), 'error')
@@ -128,12 +155,23 @@ def edit(iid):
 
 @items_bp.route('/delete/<int:iid>', methods=('POST',))
 def delete(iid):
+    # create form
     form = DeleteItemForm()
 
+    # validate form if POST request
     if form.validate_on_submit():
         item = Item.query.get(iid)
 
         if item:
+            # check that logged in user is the creator of the item
+            app_user_id = session.get('app_user_id', '')
+            if not (app_user_id and (item.app_user_id == app_user_id)):
+                flash(
+                    "You can't delete this item. "
+                    "You are not its creator", 'error')
+                return redirect(url_for('items.details', iid=iid))
+
+            # delete item
             try:
                 db.session.delete(item)
                 db.session.commit()
@@ -144,7 +182,9 @@ def delete(iid):
                 flash("Item successfully deleted", 'info')
         else:
             flash("Item with id ({}) does not exist".format(iid), 'error')
+            return redirect(url_for('categories.index'))
     else:
+        # flash form errors
         for field, errors in form.errors.items():
             for error in errors:
                 flash("{}: {}".format(field, error), 'error')
